// web/src/types/ActorsRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ActorSelection } from "./ActorSelection";
import type { InsightQuery } from "./InsightQuery";

/**
 * `POST /api/projects/{project_id}/query/actors` body.
 */
export type ActorsRequest = { query: InsightQuery, selection: ActorSelection, offset: number, limit: number, };