// web/src/types/InsightQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FunnelsQuery } from "./FunnelsQuery";
import type { LifecycleQuery } from "./LifecycleQuery";
import type { PathsQuery } from "./PathsQuery";
import type { RetentionQuery } from "./RetentionQuery";
import type { SqlQuery } from "./SqlQuery";
import type { StickinessQuery } from "./StickinessQuery";
import type { TrendsQuery } from "./TrendsQuery";

export type InsightQuery = { "kind": "TrendsQuery" } & TrendsQuery | { "kind": "FunnelsQuery" } & FunnelsQuery | { "kind": "RetentionQuery" } & RetentionQuery | { "kind": "LifecycleQuery" } & LifecycleQuery | { "kind": "StickinessQuery" } & StickinessQuery | { "kind": "PathsQuery" } & PathsQuery | { "kind": "SqlQuery" } & SqlQuery;