// web/src/types/PersonDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PersonSummary } from "./PersonSummary";

/**
 * `GET /api/projects/{project_id}/persons/{person_id}`
 */
export type PersonDetail = { person: PersonSummary, distinct_ids: Array<string>, event_count: number, first_seen: string | null, last_seen: string | null, session_count: number, };