// web/src/types/WebBreakdown.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WebBreakdownRow } from "./WebBreakdownRow";
import type { WebDimension } from "./WebDimension";

export type WebBreakdown = { dimension: WebDimension, rows: Array<WebBreakdownRow>, };